CopyManager.java tests added
